+++ src/routes/admin/post/deletePost.js
@@ -12,23 +12,25 @@ router.get('/delete-post', upload.none(), (req, res) => {
   let post_id = req.query.id;
   
   // Authenticate Admin with token and then proceed
-  AdminAuth(req, res, () => {
-    vt_posts.destroy({
-      where:{
-        id: post_id
-      }
-    })
-      .then((data) => {
-        if(data !== null) {
-          res.status(200).send({ status: 200, message: 'Post deleted!' });
-        }
-        if(data === null){
-          res.status(401).send({ status: 401, message: 'No Post found!' });
+  AdminAuth(req, res, (status) => {
+    if(status){
+      vt_posts.destroy({
+        where:{
+          id: post_id
         }
       })
-      .catch((err) => {
-        console.log(err);
-      })
+        .then((data) => {
+          if(data !== null) {
+            res.status(200).send({ status: 200, message: 'Post deleted!' });
+          }
+          if(data === null){
+            res.status(401).send({ status: 401, message: 'No Post found!' });
+          }
+        })
+        .catch((err) => {
+          console.log(err);
+        })
+    }
   })
 });
